requirements.txt
cryptography==41.0.5
pandas==2.1.2
python-decouple==3.8
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.22
